V1.0.8 beta
Engine:
✔ The README file is filled (Yaro2709);
✔ Other small changes (Yaro2709);
Mods:
✔ Fixed problems with the path of images in login @ install (Yaro2709);
✔ Fixed problems with the queue in the shipyard (Yaro2709);
✔ Fixed problems with styles in the game (Yaro2709);
✔ Fixed a language bug in the mode of referrals (SteemNova);
✔ Fixed a bug related to carrying capacity during processing (SteemNova);
✔ Fixed CleanerCronjob (SteemNova);
✔ Fixed a bug where it was possible to scan a player in vacation mode (SteemNova);
✔ Fixed a bug related to carrying capacity for espionage (SteemNova);
✔ Removed unnecessary LNG variables in the espionage report (SteemNova);
✔ Fixed removing queues from different planets in explorations (SteemNova);
✔ Fixed adding garbage after the moon battle (SteemNova);
✔ Fixed authorization error for PHP 7.4 (SteemNova);
✔ Added check for there is a request to the alliance (SteemNova);
✔ Fixed bugs related to a joint attack (SteemNova);
✔ Fixed a bug related to division by zero when sending the fleet (SteemNova);
✔ Fixed the use of deuterium to send ships from the galaxy (SteemNova);
✔ Fixed session detection for PHP7.4 (SteemNova);
✔ Fixed requests for PHP7.4 (SteemNova);
✔ Fixed a defect related to retention (SteemNova);
✔ Fixed bug when switching between planets during an interplanetary attack for PHP7.4 (SteemNova);
✔ Fixed fleet freezing when destroying satellites (SteemNova);
✔ Fixed a bug due to which the ships could move between our planets and the moon faster (SteemNova);
✔ Fixed bug in the trader (SteemNova);
✔ Fixed the vision of enemy processing missions on the alliance page (SteemNova);
✔ Fixed bugs with LNG in missions (SteemNova);
✔ Fixed a bug with the player’s bath (SteemNova);
✔ Fixed problems with time on the server (SteemNova);
✔ Fixed bugs when deleting a planet (SteemNova);
✔ Fixed erroneous display of invitations to a joint attack (SteemNova);
✔ Fixed a bug where the fleet was visible during espionage, which flew to the planet (SteemNova);
✔ Fixed bugs when deleting a planet (SteemNova);
✔ Fixed language problems between friends (SteemNova);
✔ Other small changes (Yaro2709).

+++ includes/pages/game/ShowAlliancePage.class.php
@@ -885,7 +885,7 @@ class ShowAlliancePage extends AbstractGamePage
 			}
 		}
 
-        require 'includes/classes/class.FlyingFleetHandler.php';
+        require_once 'includes/classes/class.FlyingFleetHandler.php';
 
         $available_events = array();
 
@@ -1100,7 +1100,7 @@ class ShowAlliancePage extends AbstractGamePage
 		LEFT JOIN
 			%%USERS%% AS u ON r.userId = u.id
 		INNER JOIN
-			%%STATPOINTS%% AS stat
+			%%STATPOINTS%% AS stat ON r.userId = stat.id_owner
 		LEFT JOIN
 			%%PLANETS%% AS p ON p.id = u.id_planet
 		WHERE
@@ -1138,7 +1138,7 @@ class ShowAlliancePage extends AbstractGamePage
 	protected function adminSendAnswerToApply()
 	{
 		global $LNG, $USER;
-		if(!$this->rights['SEEAPPLY'] || !$this->rights['MANAGEAPPLY']) {
+		if (!$this->rights['SEEAPPLY'] || !$this->rights['MANAGEAPPLY']) {
 			$this->redirectToHome();
 		}
 
@@ -1153,47 +1153,47 @@ class ShowAlliancePage extends AbstractGamePage
 			':applyID'	=> $applyID
 		), 'userId');
 
-		if ($answer == 'yes')
-		{
-			$sql = "DELETE FROM %%ALLIANCE_REQUEST%% WHERE applyID = :applyID";
-			$db->delete($sql, array(
-				':applyID'	=> $applyID
-			));
+		// only if alliance request still exist
+		if ($userId) {
+			if ($answer == 'yes') {
+				$sql = "DELETE FROM %%ALLIANCE_REQUEST%% WHERE applyID = :applyID";
+				$db->delete($sql, array(
+					':applyID'	=> $applyID
+				));
 
-			$sql = "UPDATE %%USERS%% SET ally_id = :allianceId, ally_register_time = :time, ally_rank_id = 0 WHERE id = :userId;";
-			$db->update($sql, array(
-				':allianceId'	=> $this->allianceData['id'],
-				':time'         => TIMESTAMP,
-				':userId'       => $userId
-			));
+				$sql = "UPDATE %%USERS%% SET ally_id = :allianceId, ally_register_time = :time, ally_rank_id = 0 WHERE id = :userId;";
+				$db->update($sql, array(
+					':allianceId'	=> $this->allianceData['id'],
+					':time'         => TIMESTAMP,
+					':userId'       => $userId
+				));
 
-			$sql = "UPDATE %%STATPOINTS%% SET id_ally = :allianceId WHERE id_owner = :userId AND stat_type = 1;";
-			$db->update($sql, array(
-				':allianceId'	=> $this->allianceData['id'],
-				':userId'       => $userId
-			));
+				$sql = "UPDATE %%STATPOINTS%% SET id_ally = :allianceId WHERE id_owner = :userId AND stat_type = 1;";
+				$db->update($sql, array(
+					':allianceId'	=> $this->allianceData['id'],
+					':userId'       => $userId
+				));
 
-			$sql = "UPDATE %%ALLIANCE%% SET ally_members = (SELECT COUNT(*) FROM %%USERS%% WHERE ally_id = :allianceId) WHERE id = :allianceId;";
-			$db->update($sql, array(
-				':allianceId'	=> $this->allianceData['id'],
-			));
+				$sql = "UPDATE %%ALLIANCE%% SET ally_members = (SELECT COUNT(*) FROM %%USERS%% WHERE ally_id = :allianceId) WHERE id = :allianceId;";
+				$db->update($sql, array(
+					':allianceId'	=> $this->allianceData['id'],
+				));
 
-			$text		= $LNG['al_hi_the_alliance'] . $this->allianceData['ally_name'] . $LNG['al_has_accepted'] . $text;
-			$subject	= $LNG['al_you_was_acceted'] . $this->allianceData['ally_name'];
+				$text		= $LNG['al_hi_the_alliance'] . $this->allianceData['ally_name'] . $LNG['al_has_accepted'] . $text;
+				$subject	= $LNG['al_you_was_acceted'] . $this->allianceData['ally_name'];
+			} else {
+				$sql = "DELETE FROM %%ALLIANCE_REQUEST%% WHERE applyID = :applyID";
+				$db->delete($sql, array(
+					':applyID'	=> $applyID
+				));
+
+				$text		= $LNG['al_hi_the_alliance'] . $this->allianceData['ally_name'] . $LNG['al_has_declined'] . $text;
+				$subject	= $LNG['al_you_was_declined'] . $this->allianceData['ally_name'];
+			}
+
+			$senderName	= $LNG['al_the_alliance'] . $this->allianceData['ally_name'] . ' [' . $this->allianceData['ally_tag'] . ']';
+			PlayerUtil::sendMessage($userId, $USER['id'], $senderName, 2, $subject, $text, TIMESTAMP);
 		}
-		else
-		{
-			$sql = "DELETE FROM %%ALLIANCE_REQUEST%% WHERE applyID = :applyID";
-			$db->delete($sql, array(
-				':applyID'	=> $applyID
-			));
-
-			$text		= $LNG['al_hi_the_alliance'] . $this->allianceData['ally_name'] . $LNG['al_has_declined'] . $text;
-			$subject	= $LNG['al_you_was_declined'] . $this->allianceData['ally_name'];
-		}
-
-		$senderName	= $LNG['al_the_alliance'] . $this->allianceData['ally_name'] . ' ['.$this->allianceData['ally_tag'].']';
-		PlayerUtil::sendMessage($userId, $USER['id'], $senderName, 2, $subject, $text, TIMESTAMP);
 		$this->redirectTo('game.php?page=alliance&mode=admin&action=mangeApply');
 	}
